before adding swagger

+++ src/tasks/tasks.service.ts
@@ -1,18 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, Req } from '@nestjs/common';
 import { TaskStatus } from './task-status.enum';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
+import { GetUser } from 'src/auth/get-user.decorator';
+import { User } from 'src/auth/user-db/user.entity';
 @Injectable()
 export class TasksService {
     constructor(
         @InjectRepository(TaskRepository)
         private taskRepository: TaskRepository
     ) { }
-    async getTask(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTask(filterDto: GetTaskFilterDto,
+        user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
     // getAllTasks(): Task[] {
     //     return this.tasks;
@@ -34,8 +37,8 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`Task with id : ${id} not found-غير موجود-`);
@@ -61,8 +64,11 @@ export class TasksService {
     // //     this.tasks.push(task);
     // //     return task;
     // // }
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto,
+        user: User,
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user
+        );
         // const { title, description } = createTaskDto;
         // const task = new Task();
         // task.title = title;
@@ -82,8 +88,8 @@ export class TasksService {
     //     this.tasks.push(task);
     //     return task;
     // }
-    async updateTask(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTask(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
@@ -93,8 +99,8 @@ export class TasksService {
     //     task.status = status;
     //     return task;
     // }
-    async deleteTask(id: number): Promise<any> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<any> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         if (result.affected === 0) {
             throw new NotFoundException(`Task with id : ${id} not found-غير موجود-`);
         } else {
